cd system: ejected changelog processing
